feat: add "rules" concept, `resolve.root` name-clash detection with node_modules packages, file path existence checks (#81)
BREAKING CHANGE: validate() now takes only two arguments. The schema which could be passed as second argument can now be supplied via property `schema` on the options object which is now the second argument. The now preferred way to supply extra properties to the schema is supplying a a Joi.object via the options property `schemaExtension`, that *just contains the properties you want to add* (see README under "Customizing"). The schema used by this library **is not exported anymore**, as we now parameterize it by supplying "rules" which toggle / modify parts of the schema (Have a look at `src/index.js` to learn more).