feat: Allow performance.hints in webpack config (#150)
* feat: Allow performance.hints in webpack config

Webpack 2.1.0-beta.28 adds `performance.hints` (which take boolean
values) as a way to enable/disable performance warnings. (The first
performance warning is about bundle size.)

* fix: Fix name of an import

The test file for the `performance` property should import the test utils module using `utils` in

lowercase but instead imported `Utils` in uppercase. This worked fine on my machine because my

filesystem is case-insensitive but fails in Travis because its filesystem is case-sensitive.